momentum-scanner/requirements.txt
yfinance==0.2.37
pandas==2.2.2
numpy==1.26.4
scipy==1.13.0
ta==0.11.0
schedule==1.2.1
pytz==2024.1
python-dotenv==1.0.1
requests==2.32.3
pyarrow==16.0.0
